Refactor Jog and Preferences functionality into separate file
This is step 1 towards cleaning up the 3000 line psng.py file :)
