+++ .gitignore
@@ -1,2 +1,3 @@
 *.pyc
+*.swp
 .vscode
