+++ psng/install_add_to_your.ini
@@ -1,7 +1,7 @@
 [DISPLAY]
 DISPLAY = axis
 EMBED_TAB_NAME=PSNG
-EMBED_TAB_COMMAND=gladevcp -c gladevcp -x {XID} -u psng/psng.py -H psng/psng.hal psng/psng.glade
+EMBED_TAB_COMMAND=gladevcp -c probe -x {XID} -u psng/python -H psng/psng.hal psng/psng.glade
 
 [TOOLSENSOR]
 # Absolute coordinates of the toolsetter pad
